+++ src/parser/sampling/RandomWalkSampler.java
@@ -35,7 +35,7 @@ public class RandomWalkSampler {
 		predInst.heads = new int[len];
 		predInst.deplbids = new int[len];
         
-        //double[] score = new double[len];
+        //float[] score = new float[len];
         //int[] depList = new int[len];
         //int size = 0;
 
@@ -55,7 +55,7 @@ public class RandomWalkSampler {
     			//	if (candH == curr || lfd.isPruned(candH, curr))
     			//		continue;
     			//	
-    			//	//double s = lfd.getArcScore(candH, curr);
+    			//	//float s = lfd.getArcScore(candH, curr);
     			//	
     			//	//if (addLoss) {
     			//	//	  if (candH != inst.heads[curr])
@@ -94,7 +94,7 @@ public class RandomWalkSampler {
 		predInst.heads = new int[len];
 		predInst.deplbids = new int[len];
         
-        double[] score = new double[len];
+        float[] score = new float[len];
         int[] depList = new int[len];
         int size = 0;
 
@@ -114,7 +114,7 @@ public class RandomWalkSampler {
     				if (candH == curr || lfd.isPruned(candH, curr))
     					continue;
     				
-    				double s = lfd.getArcScore(candH, curr);
+    				float s = lfd.getArcScore(candH, curr);
     				
     				if (addLoss) {
     					  if (candH != inst.heads[curr])
@@ -139,13 +139,13 @@ public class RandomWalkSampler {
     	return predInst;
     }
     
-    private int samplePoint(double[] score, int N, Random r) {
-    	double sumScore = Double.NEGATIVE_INFINITY;
+    private int samplePoint(float[] score, int N, Random r) {
+    	float sumScore = Float.NEGATIVE_INFINITY;
     	for (int i = 0; i < N; i++) {
     		sumScore = Utils.logSumExp(sumScore, score[i]);
     	}
-    	double logp = Math.log(r.nextDouble() + 1e-60);
-    	double cur = Double.NEGATIVE_INFINITY;
+    	float logp = (float)Math.log(r.nextFloat() + 1e-60);
+    	float cur = Float.NEGATIVE_INFINITY;
     	int ret = 0;
     	for (; ret < N; ret++) {
     		cur = Utils.logSumExp(cur, score[ret]);
@@ -163,7 +163,7 @@ public class RandomWalkSampler {
 //                if (v == 0 || u == v || lfd.isPruned(u, v))
 //                    System.out.print("0.00\t");
 //                else {
-//                    double s = lfd.getArcScore(u, v);
+//                    float s = lfd.getArcScore(u, v);
 //                    //int l = options.learnLabel ? staticTypes[u][v] : 0;
 //                    //s += options.learnLabel ? lfd.getLabeledArcScore(u, v, l) : 0.0;
 //                    if (addLoss) {
